requirements.txt
Django==4.2
gradio==3.11.0
huggingface-hub==0.22.2
djangorestframework==3.14.0
